backend_automations/pyproject.toml
[tool.poetry]
name = "backend_automations"
version = "0.1.0"
description = ""
authors = ["Rémi - BESNARD CONSULTING <remi@besnard-consulting.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.3"
django-environ = "^0.4.5"
gunicorn = "^20.1.0"
requests = "^2.25.1"
psycopg2-binary = "^2.8.6"
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]
mypy = "^0.812"
black = "^21.5b1"
pytest-django = "^4.2.0"
factory-boy = "^3.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
